Add first draft of intro section 1 (what is IPFS)
Also modifies how sections are sorted so we can get these in order.

Adds some bits to the theme in a temporary, hacky way. @lgierth, I couldn't find the repo for this theme that is mentioned in its TOML file; where does it live/how are we updating/editing it?